Fix sass compilation/compression issues
