+++ Gemfile
@@ -11,18 +11,17 @@ gem 'delayed_job_active_record'
 
 # Metadata presenter - if you need to be on development you can uncomment
 # one of these lines:
-gem 'metadata_presenter',
-    github: 'ministryofjustice/fb-metadata-presenter',
-    branch: 'editable-content-web-component'
+# gem 'metadata_presenter',
+#     github: 'ministryofjustice/fb-metadata-presenter',
+#     branch: 'editable-content-web-component'
 # gem 'metadata_presenter', path: '../fb-metadata-presenter'
-# gem 'metadata_presenter', '3.0'
+gem 'metadata_presenter', '3.0'
 
 gem 'aws-sdk-s3'
 gem 'aws-sdk-sesv2'
 gem 'faraday'
 gem 'faraday_middleware'
 gem 'fb-jwt-auth', '0.10.0'
-gem 'govspeak', '~> 7.1'
 gem 'govuk-components'
 gem 'govuk_design_system_formbuilder'
 gem 'hashie'
