+++ app/javascript/styles/shared/_content.scss
@@ -18,32 +18,4 @@
 // Apply GOVUk link styles to all links in editable content areas
 .fb-editable[data-fb-content-type="content"] a {
   @extend %govuk-link;
-
-    table {
-    @include govuk-font($size: 19);
-    @include govuk-text-colour;
-    width: 100%;
-    @include govuk-responsive-margin(6, "bottom");
-
-    border-spacing: 0;
-    border-collapse: collapse;
-  }
-
-  thead td,
-  thead th {
-    @include govuk-typography-weight-bold;
-  }
-
-  td, th {
-    padding: govuk-spacing(2) govuk-spacing(4) govuk-spacing(2) 0;
-    border-bottom: 1px solid $govuk-border-colour;
-    text-align: left;
-    vertical-align: top;
-  }
-
-  td:last-child,
-  th:last-child {
-    padding-right: 0;
-  }
-
 }
