+++ config/environments/development.rb
@@ -1,10 +1,7 @@
 require "active_support/core_ext/integer/time"
-# require "./lib/skipping_sass_compressor"
 
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
-
-  # config.assets.css_compressor = SkippingSassCompressor.new
 
   # In the development environment your application's code is reloaded any time
   # it changes. This slows down response time but is perfect for development
