--- lib/skipping_sass_compressor.rb
@@ -1,11 +1,0 @@
-class SkippingSassCompressor
-  def compress(string)
-    options = { syntax: :scss, cache: false, read_cache: false, style: :compressed}
-    begin
-      Sprockets::Autoload::SassC::Engine.new(string, options).render
-    rescue => e
-      puts "Could not compress '#{string[0..65]}'...: #{e.message}, skipping compression"
-      string
-    end
-  end
-end
